requirements.txt
# Core
python-dotenv>=1.0
backoff>=2.2
requests>=2.32
httpx>=0.27
# LLM (Typhoon API)
openai>=1.35
# Search (Tavily + SerperAPI + JINA - NO SERPAPI)
tavily-python>=0.2.9
# Note: SerperAPI uses simple requests (no library needed)
# Note: JINA_API uses simple requests (no library needed)
# Front-end demo
streamlit>=1.35
# Data processing
pandas>=1.5.0
# PowerPoint generation
python-pptx>=1.0.0
# PDF generation
reportlab>=4.0.0
# API endpoints
flask>=2.3.0
werkzeug>=2.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
# Azure Storage
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
# AI Image Generation & Gemini API
google-generativeai>=0.8.0
google-genai>=0.3.0
# Pillow for image processing
Pillow>=10.0.0
# Beautiful Soup for HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0
# PostgreSQL Database
psycopg2-binary>=2.9.9
# Production server
gunicorn>=21.0.0
